Final attempts
Youtube functionality researched and added. Not much was changed as could not get the final commits to work. But did create a small database search to cross reference with youtube. Should search both at the same time and return a recipe and a video as well. In a seperate commit the research log will be updated. As for time on upload, I left for thanksgiving week without committing and I apologize.